split with client and interface
